+++ libsolidity/ast/ASTAnnotations.h
@@ -138,6 +138,7 @@ struct InlineAssemblyAnnotation: StatementAnnotation
 		Declaration const* declaration = nullptr;
 		bool isSlot = false; ///< Whether the storage slot of a variable is queried.
 		bool isOffset = false; ///< Whether the intra-slot offset of a storage variable is queried.
+		std::optional<ASTString> enumValue = {}; ///< The name of the enum value, if one is queried.
 		size_t valueSize = size_t(-1);
 	};